+++ public/css/styles.css
@@ -23,7 +23,8 @@ body {
     width: 220px;
     border: 1px solid #000;
     padding: 12px;
-    box-sizing: border-box; }
+    box-sizing: border-box;
+    margin-bottom: 15px; }
     .gallery__img-block img {
       width: 196px; }
     .gallery__img-block:hover {
@@ -96,5 +97,24 @@ nav {
   color: #fff;
   min-height: 80px;
   background-color: #555; }
+
+.main-menu__link {
+  display: inline-block;
+  padding: 8px 12px;
+  margin: 0 5px;
+  color: #222;
+  background-color: #eee;
+  border: 1px solid #444;
+  text-transform: uppercase;
+  text-decoration: none; }
+  .main-menu__link:hover {
+    color: #eee;
+    background-color: #222;
+    border: 1px solid #ddd; }
+
+.message {
+  background-color: #eee;
+  color: #222;
+  padding: 8px 12px; }
 
 /*# sourceMappingURL=styles.css.map */
